requirements.txt
requests==2.32.3
Django~=4.2.16
djangorestframework==3.15.2
psycopg2-binary>=2.9.0
python-dotenv==1.0.1